调整漫画详情页元数据区布局：R18 chip 独立固定行，概要移至阅读按钮下方。
