+++ app/lib/ui/core/layout/detail_meta_chip_row_layout.dart
@@ -1,0 +1,2 @@
+/// 与 [OutlinedMetaChip] / [R18RatingChip] 纵向占位一致，用于详情页 chip 行固定高度。
+const double kDetailMetaChipRowHeight = 28;
