+++ app/lib/ui/features/library/views/comic_detail_page/widgets/comic_detail.dart
@@ -81,6 +81,7 @@ class ComicDetail extends StatelessWidget {
           ),
           ComicDetailSummaryMetaRow(comic: comic),
           ComicDetailPrimaryActions(comic: comic),
+          ComicDetailDescription(comic: comic),
         ],
       ),
     );
